Desktop/grape ai chat test/Potato-Local-AI-main/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
cairosvg
tkinterweb
brotli==1.2.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
ddgs==9.10.0
duckduckgo_search==8.1.1
fake-useragent==2.2.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
lxml==6.0.2
ollama==0.6.1
pillow==12.1.1
primp==1.0.0
pydantic==2.12.5
pydantic_core==2.41.5
requests==2.32.5
socksio==1.0.0
soupsieve==2.8.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
